archive components from launch page containers
