+++ packages/ui/src/archive/modals/SkillsModal/SkillsModal.stories.tsx
@@ -1,22 +1,27 @@
 import { MockedProvider } from "@apollo/client/testing";
 import { FIND_ALL_CATEGORIES } from "@eden/package-graphql";
+import { getSkillRoleTypeMockArray } from "@eden/package-mock";
 import { ComponentMeta, ComponentStory } from "@storybook/react";
 
-import { SkillsSubcategoryModal } from "./SkillsSubcategoryModal";
+import { SkillsModal } from "./SkillsModal";
 
 export default {
-  title: "Modals/SkillsSubcategoryModal",
-  component: SkillsSubcategoryModal,
+  title: "Archive/Modals/SkillsModal",
+  component: SkillsModal,
   argTypes: {},
-} as ComponentMeta<typeof SkillsSubcategoryModal>;
+} as ComponentMeta<typeof SkillsModal>;
 
-const Template: ComponentStory<typeof SkillsSubcategoryModal> = (args) => (
-  <SkillsSubcategoryModal {...args} />
+const Template: ComponentStory<typeof SkillsModal> = (args) => (
+  <SkillsModal {...args} />
 );
 
 export const Default = Template.bind({});
 Default.args = {
+  skills: getSkillRoleTypeMockArray(9),
   isOpen: true,
+  handelAddSkills() {
+    console.log("SkillsModal HandelAddSkills");
+  },
 };
 
 Default.parameters = {
