+++ packages/ui/src/cards/project/index.ts
@@ -1,3 +1,2 @@
 export * from "./ProjectEditSelectorCard";
-export * from "./ProjectLayoutCard";
 export * from "./ProjectNodeMatchCard";
