+++ packages/ui/src/modals/index.ts
@@ -14,18 +14,10 @@ export * from "./ProjectNodeMatchModal";
 export * from "./RequirementsModal";
 export * from "./ReviewModal";
 export * from "./RoleDescriptionModal";
-export * from "./RoleModal";
 export * from "./SavingProjectModal";
 export * from "./SelectNodesModal";
 export * from "./SendMessageToUserModal";
-export * from "./ShortlistMemeberModal";
-export * from "./SkillMatchModal";
-export * from "./SkillsCategoryModal";
-export * from "./SkillsModal";
-export * from "./SkillsOnCategoryModal";
-export * from "./SkillsSubcategoryModal";
 export * from "./SkipFlowModal";
-export * from "./SocialMediaModel";
 export * from "./StaticModal";
 export * from "./UserDiscoverModal";
 export * from "./UserInviteModal";
